+++ jupyter_server/auth/__init__.py
@@ -1,1 +1,3 @@
+from .authorizer import *  # noqa
+from .decorator import authorized  # noqa
 from .security import passwd  # noqa
