+++ jupyter_server/files/handlers.py
@@ -7,8 +7,8 @@ from base64 import decodebytes
 
 from tornado import web
 
+from jupyter_server.auth import authorized
 from jupyter_server.base.handlers import JupyterHandler
-from jupyter_server.services.auth.decorator import authorized
 from jupyter_server.utils import ensure_async
