+++ jupyter_server/kernelspecs/handlers.py
@@ -2,7 +2,7 @@ from tornado import web
 
 from ..base.handlers import JupyterHandler
 from ..services.kernelspecs.handlers import kernel_name_regex
-from jupyter_server.services.auth.decorator import authorized
+from jupyter_server.auth import authorized
 
 
 AUTH_RESOURCE = "kernelspecs"
