+++ jupyter_server/services/api/handlers.py
@@ -10,7 +10,7 @@ from ...base.handlers import APIHandler
 from ...base.handlers import JupyterHandler
 from jupyter_server._tz import isoformat
 from jupyter_server._tz import utcfromtimestamp
-from jupyter_server.services.auth.decorator import authorized
+from jupyter_server.auth import authorized
 from jupyter_server.utils import ensure_async
