--- jupyter_server/services/auth/__init__.py
@@ -1,1 +1,0 @@
-from .decorator import authorized  # noqa
