+++ jupyter_server/services/kernels/handlers.py
@@ -26,7 +26,7 @@ from ...base.zmqhandlers import deserialize_binary_message
 from jupyter_server.utils import ensure_async
 from jupyter_server.utils import url_escape
 from jupyter_server.utils import url_path_join
-from jupyter_server.services.auth.decorator import authorized
+from jupyter_server.auth import authorized
 
 
 AUTH_RESOURCE = "kernels"
