+++ jupyter_server/services/nbconvert/handlers.py
@@ -5,7 +5,7 @@ from anyio.to_thread import run_sync
 from tornado import web
 
 from ...base.handlers import APIHandler
-from jupyter_server.services.auth.decorator import authorized
+from jupyter_server.auth import authorized
 
 
 AUTH_RESOURCE = "nbconvert"
