+++ jupyter_server/services/security/handlers.py
@@ -5,7 +5,7 @@ from tornado import web
 
 from . import csp_report_uri
 from ...base.handlers import APIHandler
-from jupyter_server.services.auth.decorator import authorized
+from jupyter_server.auth import authorized
 
 
 AUTH_RESOURCE = "csp"
