+++ jupyter_server/services/sessions/handlers.py
@@ -17,7 +17,7 @@ from tornado import web
 from ...base.handlers import APIHandler
 from jupyter_server.utils import ensure_async
 from jupyter_server.utils import url_path_join
-from jupyter_server.services.auth.decorator import authorized
+from jupyter_server.auth import authorized
 
 
 AUTH_RESOURCE = "sessions"
